tapir-openapi-jsoniter: make it a real tapir module
Rename openapi-jsoniter -> tapir-openapi-jsoniter (dir, artifact
kindlings-tapir-openapi-jsoniter, package hearth.kindlings.tapiropenapijsoniter).

Promote tapir-openapi-docs from a test-only dep to a production compile
dependency on JVM + Scala.js (it is not published for Native), compiled from an
extra src/main/scala-tapir source directory. The pure jsoniter codecs for the
sttp-apispec OpenAPI model stay cross-platform (incl. Native).

Add the TapirOpenApi bridge: tapir endpoints -> sttp-apispec OpenAPI -> JSON via
this module's circe-free jsoniter codec (toOpenAPI/toJson/endpointsToJson). This
is the module's headline feature it was missing -- serialize tapir-generated
OpenAPI without pulling in circe (unlike sttp-apispec openapi-circe).

TapirInteropJvmSpec now drives the bridge end-to-end and still cross-checks
byte-for-byte against circe. 122 tests green on JVM 2.13+3; bridge compiles on
JS 2.13+3; codecs compile on Native.
