+++ tapir-openapi-jsoniter/src/test/resources/extending-recursive.json
