+++ tapir-openapi-jsoniter/src/test/resources/petstore/basic-petstore.json
