+++ tapir-openapi-jsoniter/src/test/resources/petstore/header-petstore.json
