+++ tapir-openapi-jsoniter/src/test/resources/securityScheme/security-scheme-with-scopes.json
