+++ tapir-openapi-jsoniter/src/test/resources/self-describing-schema.json
