+++ tapir-openapi-jsoniter/src/test/resources/spec/3.0/schema.json
