+++ tapir-openapi-jsoniter/src/test/resources/spec/3.1/any_and_nothing2.json
