+++ tapir-openapi-jsoniter/src/test/resources/spec/3.1/schema.json
